Move configuration sources into subdirectory.
