+++ include/bitcoin/server.hpp
@@ -15,17 +15,17 @@
  */
 
 #include <bitcoin/node.hpp>
-#include <bitcoin/server/config.hpp>
 #include <bitcoin/server/define.hpp>
 #include <bitcoin/server/dispatch.hpp>
-#include <bitcoin/server/endpoint.hpp>
 #include <bitcoin/server/message.hpp>
 #include <bitcoin/server/publisher.hpp>
 #include <bitcoin/server/server_node.hpp>
-#include <bitcoin/server/settings.hpp>
 #include <bitcoin/server/subscribe_manager.hpp>
 #include <bitcoin/server/version.hpp>
 #include <bitcoin/server/worker.hpp>
+#include <bitcoin/server/config/config.hpp>
+#include <bitcoin/server/config/endpoint.hpp>
+#include <bitcoin/server/config/settings.hpp>
 #include <bitcoin/server/service/blockchain.hpp>
 #include <bitcoin/server/service/compat.hpp>
 #include <bitcoin/server/service/fetch_x.hpp>
