+++ src/config/endpoint.cpp
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-#include <bitcoin/server/endpoint.hpp>
+#include <bitcoin/server/config/endpoint.hpp>
 
 #include <cstdint>
 #include <iostream>
