+++ src/config/settings.cpp
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-#include <bitcoin/server/settings.hpp>
+#include <bitcoin/server/config/settings.hpp>
 
 #include <iostream>
 #include <string>
