+++ src/dispatch.cpp
@@ -25,12 +25,12 @@
 #include <boost/filesystem.hpp>
 #include <boost/program_options.hpp>
 #include <bitcoin/node.hpp>
-#include <bitcoin/server/config.hpp>
+#include <bitcoin/server/config/config.hpp>
+#include <bitcoin/server/config/settings.hpp>
 #include <bitcoin/server/message.hpp>
 #include <bitcoin/server/server_node.hpp>
 #include <bitcoin/server/publisher.hpp>
 #include <bitcoin/server/subscribe_manager.hpp>
-#include <bitcoin/server/settings.hpp>
 #include <bitcoin/server/version.hpp>
 #include <bitcoin/server/worker.hpp>
 #include <bitcoin/server/server_node.hpp>
